Migrate old TNNT backend design discussions into code and .md file
Similar to the .md file I just created in the TNNT game repository, this
does the same thing for design discussions pertaining to the backend and
the tournament structure, allowing me to clear out the forum where the
stagnant design discussions actually took place.

At present, several of the new additions with a more technical bent are
put in relevant places in files as comments and TODOs rather than in
old_design_discussions; if the comments (but not the TODOs) turn out to
be poorly findable, they should probably be moved into
old_design_discussions.