Mention in the readme how to poll a local xlogfile
I intend to do this every year when setting up my dev environment
(because I don't want to download the full last year xlogfiles from
hardfought), and forget how to every year until I go poking around in
the source. Document how it works in the readme so it's more obvious
that it's an option.