Don't show zeroes on leaderboards
Resolving a TODO from last year. The design notes enumerated a bunch of
leaderboards where zero entries are redundant. I noticed today that
every leaderboard not in that set can't show zeroes by definition
anyway, so it was pretty easy to implement a broad prevention on zero
entries.

My previous notes also mentioned that there's one that's
questionable/arguable, the max score in a game leaderboard; but I don't
expect anyone to get upset that all their games have resulted in 0
points and they don't appear on the leaderboard. It's pretty easy to get
points in NetHack.

Also in this commit, I expanded a couple leaderboard titles to clarify
they're about ascensions and removed a redundant element from the
splats board definition.