Remove vanilla achievement tracking code
I added the vanilla achievements to the list of regular TNNT
achievements in the TNNT source code, so this is the other half of it.
We no longer need a manually maintained vanilla_achievements.yaml that
is tacked on to the achievements fixture; the generate-from-TNNT
automatic step is all that's needed. The scoreboard now doesn't read
anything from the "achieve" field in the xlogfile.

Because I kept their place at the very start of the achievement list,
regenerating achievements.yaml involved updating all the bit positions
of all the other achievements. Not the last time this will be done for
2025, as we add more new achievements.