Create 'list of all ascensions' page that is accessed via the stats page.
This list template is based off of the the 'last 10 ascensions' block as
seen on the main page. All columns are sortable. When testing out the
sorting, found and fixed a couple bugs with sorting by date and
duration.

There's more I think we can do with this 'all ascensions' page, but this
is a good start.