Fix: date/time formatting.
When TNNT first began, that date/time was displayed in ISO format, just
like every other NetHack website that tracks online games and displays
the xlogfiles. It was like this for the first three tournaments. Then in
2021 we completely reworked the website using Django, and it's
default date/time format is not ISO, but more like 'Nov. 10th, 4:20
a.m.' which I never liked. Trying to sort using that format is difficult
and the sorting does not always work. Switching to using an ISO format
1) makes sorting much simpler to work with and 2) looks better, and
harkens back to the way it was originally displayed.