Fix: various data formatting, add UTC to endtime column headers.
First bit - add (UTC) next to 'endtime' and similar column headers so
there's no mistaking the time zone TNNT uses.

Second bit - various tables on player and clan pages use flexbox
styling, which if not set properly will cause like tables with less rows
to 'stretch' vertically to attempt to match the height of the table with
the most rows. This caused a disproportinate alignment between 'general
information', and 'trophies' tables on player pages (and 'members' table
on clan pages). Been wanting to fix this for a long time... align-items
in the right block in the global .css file saves the day.