Fix: linking to non-existent dumplogs of startscummed games.
Part way through the 2024 tournament, a patch was added to the game so
it would not produce dumplog files if the game were startscummed (quit
or escaped within 100 turns or less). The TNNT scoreboard still had
links to the dumplog that would have been generated, but was not. This
commit removes web links for startscummed games.