Fix: pollxlogs manual import couldn't detect server origin.
When in production mode and 'manage.py pollxlogs' is ran, it downloads
the xlogfiles from each server and processes each one knowing its server
of origin, which in turn allows the scoreboard to create the proper
dumplog file links. Manually running the same command and specifying a
file would skip that check, and treat every xlogfile entry as coming
from the US server. This commit allows manually running this process to
scan for and utilize server information in the xlogfile being scanned,
which will produce correct dumplog file links.