Add a "no multiple accounts" rule to the FAQ
This rule is intended to address some of the abuses we saw in the 2024
tournament, most of which relied on people using sockpuppet accounts to
run items around between swap chests while temporarily parking their
main account that they actually wanted to play on. That's not really in
the spirit of the tournament. Hence, this rule. You have 1 account, you
use that account to play NetHack.

This rule does have an exception for round-robin accounts (which are a
well-defined sort of account sharing), and it doesn't apply to playing
multiple games on the different Hardfought servers under the same
account (it's still one account).