Add debug mode disclaimer on main page, clan management login page.
Currently the site is in debug mode for testing, a player tried to login
to the clan management page and was confused as to why they couldn't
login. Figured it'd be good to put up a disclaimer, both on the login
page as well as the main index page, to hopefully avoid any
misunderstandings.