Fix: clan trophy tracker not updating when players joined/left clan.
Technically it would update, but not until the next aggregate session
was ran, which for the tournament is every 10 minutes. Now, the clan
trophy tracker will update immediately when a player joins or leaves the
clan for whatever reason.

Individual player trophy trackers will still update with the aggregate
cycle.