Avoid Unicode characters in template source code
This sort of thing works up until it doesn't. Keep the template source
healthy by replacing non-ascii characters in the templates with HTML
entities which a browser will render to the same thing.

As part of this, I changed the "complete mines and sokoban" symbol in
the trophy tracker from an unclear green check mark (unclear because
this tracker is also used for ascensions, and the check mark is easily
confused as "you did this one!") to a pick-axe and rock emoji. I'm
typically iffy on emoji and it's possible people will think this looks
too goofy for the TNNT scoreboard. If so, we can change it to something
else, but not back to a green check.