Move logged-in clan trophy tracker to single-clan page
In 413da0e, a tracker for clan trophies was added to the clan management
page as well as the singleplayerorclan template. However, it had some
drawbacks: the code was mostly duplicated in the two templates, and
design-wise, I didn't want the clan admin page to be contaminated with
tournament stats when the natural place for clan stats would be the clan
results page.

I was in favor of just showing the tracker on the clan page, but K2
mentioned that it's partly an incentive to get people to log into the
site, and partly a speed bump for anyone comparing their progress
towards trophies against other clans - though I'm still unclear on why
seeing other clans' trophy trackers would be an advantage, it's not like
trophies are worth points.

Regardless, this change keeps the clan tracker private and locked behind
a login, by making it only show on your own clan page when you are
logged in. This gets rid of the code duplication and also resolves a
bit of inefficiency since the single clan view was computing and passing
trophy_grid to the template, which didn't render it. The correct Django
pattern for cases like this is for the view to decide if data needs to
be sent, rather than sending it anyway and letting the template decide.