Fix: put noscript warning on clan disband button
Someone noted that with JavaScript disabled, the button to disband a
clan is still unprotected and dangerous. Short of implementing an entire
page warning the person the action can't be undone, there's no way to
protect this button, but this commit adds a text warning that shows up
only when JavaScript is off.