Make the clan trophy progress tracker visible to all.
This commit makes it so each clan's trophy progress tracker is visible
to the public. The trophy status summary has also been improved: it
shows all of the remaining trophies that weren't a part of this
initially, changes the display from one long column to two shorter
columns, and will list every conduct earned/not earned for easier
tracking.