Overall statistics overhaul, achievements link now on the overall header.
We featured an overall statistics page, but also had some of that data
on the main index page. Basically an unnecessary duplication of code.
This commit removes the /stats page entirely, and moves all of the data
featured from that page into the overall statistics block on the main
index page, including the links to the other statistics pages.

To fill in the gap left by the stats link being removed from the header,
I added a link to the main achievements page. Players have been asking
for a persistent link to this for years, figured we were overdue on
providing one.