Fix: don't limit API calls from localhost; return all clans/players.
Was figuring out why the TNNT channel bot was not reporting achievements
and such for some players, come to find out the API would only return
the top 50 players and top 20 clans... oops.

So when that was fixed, the channel bot was being rate-limited calling
the REST API. The limit is 100 calls per minute per source - localhost
is now excluded from that limit.